feature!: game can handle states and use of PopulateCellsAray
The 3 states are:
FIRST_CLICK - waits for the first click to distribute
the mines and guarantee that the first opened cell is not a mine;
PLAYING - after the first click and while the player does not click on a mine cell.
GAME_OVER - if the cell clicked is a mine, then it's a Game Over
situation. Mouse inputs are not accepted anymore until the game is restarted.